backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
birdnet==0.2.11
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.4
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
idna==3.11
kagglehub==0.4.3
kagglesdk==0.1.15
keras==3.13.2
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.2
opt_einsum==3.4.0
optree==0.18.0
ordered-set==4.1.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
protobuf<6.0
psutil==7.2.2
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
scipy==1.17.0
setuptools==81.0.0
six==1.17.0
soundfile==0.13.1
SQLAlchemy==2.0.46
sqlmodel==0.0.32
starlette==0.52.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.1.1
PyJWT[crypto]==2.10.1
google-cloud-storage==2.18.2
google-cloud-run==0.10.16
psycopg2-binary==2.9.10
httpx==0.28.1
